Committing a bunch of crapola. Don't remember it all, but will notate this with most recent change. Will be committing more granularly as time goes on.
Most recent change is a POC between the DistributionFitter and MarkovComponent. We will need to use something different to evaluate integrals other than quad, or find a way to make QUAD more accurate. Currently, the Kelly criterion is yielding no solution to a fitted distribution that should have a solution, as expectation is positive. However, when calculating expectation with the quad function over a large bounds, we get NEGATIVE. BUT! for this same integral, separated into two (negative + positive bounds), we get a POSITIVE expectation. This is highly irregular and should not occur with an accurate integration function.

TODO: find out why quad is yielding incorrect values for larger bounds, and determine fix.